get items from backend and list them
