+++ packages/client/src/components/home.component.tsx
@@ -1,4 +1,4 @@
-import AddItemDrawer from "./add-item-drawer.component"
+import AddItemDrawer from "./drawer/add-item-drawer.component"
 import Inventory from "./grid/inventory.component"
 
 export default function Home() {
